+++ 0-Coins/app/src/main/java/com/axiasoft/android/zerocoins/ui/features/available_books/domain/apis/BitsoOrderBooksApi.kt
@@ -7,7 +7,7 @@ import com.axiasoft.android.zerocoins.network.bitso.BitsoApiPaths
 import com.axiasoft.android.zerocoins.network.bitso.models.BitsoBaseResponse
 import com.axiasoft.android.zerocoins.network.connections.HttpConnectionManager
 import com.axiasoft.android.zerocoins.ui.features.available_books.domain.models.data.exchange_order_book.response.ExchangeOrderBookResponse
-import com.axiasoft.android.zerocoins.ui.features.available_books.domain.models.data.order_book.response.ListOrderBookResponse
+import com.axiasoft.android.zerocoins.ui.features.available_books.domain.models.data.open_orders_book.response.ListOrderBookResponse
 import com.axiasoft.android.zerocoins.ui.features.available_books.domain.models.data.ticker.response.Ticker
 import retrofit2.http.GET
 import retrofit2.http.Query
