+++ 0-Coins/app/src/main/java/com/axiasoft/android/zerocoins/ui/features/available_books/domain/models/data/open_orders_book/response/Bids.kt
@@ -1,8 +1,8 @@
-package com.axiasoft.android.zerocoins.ui.features.available_books.domain.models.data.order_book.response
+package com.axiasoft.android.zerocoins.ui.features.available_books.domain.models.data.open_orders_book.response
 
 import com.google.gson.annotations.SerializedName
 
-data class Ask (
+data class Bids(
     @SerializedName("book"  ) var book          : String? = null,
     @SerializedName("price" ) var price         : String? = null,
     @SerializedName("amount") var amount        : String? = null,
