+++ 0-Coins/app/src/main/java/com/axiasoft/android/zerocoins/ui/features/available_books/views/ui_states/ListOrderBookScreenState.kt
@@ -1,7 +1,7 @@
 package com.axiasoft.android.zerocoins.ui.features.available_books.views.ui_states
 
-import com.axiasoft.android.zerocoins.ui.features.available_books.domain.models.data.order_book.response.Ask
-import com.axiasoft.android.zerocoins.ui.features.available_books.domain.models.data.order_book.response.Bids
+import com.axiasoft.android.zerocoins.ui.features.available_books.domain.models.data.open_orders_book.response.Ask
+import com.axiasoft.android.zerocoins.ui.features.available_books.domain.models.data.open_orders_book.response.Bids
 
 sealed class ListOrderBookScreenState {
     data class Success(val asks: ArrayList<Ask>, val bids: ArrayList<Bids>) : ListOrderBookScreenState()
